# docs/api_spec.yaml
openapi: "3.0.3"
info:
  title: Distributed Synchronization System API
  description: |
    API untuk Distributed Synchronization System yang mengimplementasikan:
    - Distributed Lock Manager (Raft Consensus + Deadlock Detection)
    - Distributed Queue (Consistent Hashing + At-least-once delivery)
    - MESI Cache Coherence Protocol (LRU replacement)
    - PBFT Byzantine Fault Tolerance (Bonus A)
    - JWT Authentication + RBAC + Audit Logging (Bonus D)
  version: "1.0.0"
  contact:
    name: Distributed Systems - Tugas 2
servers:
  - url: http://localhost:8001
    description: Node 1 (default)
  - url: http://localhost:8002
    description: Node 2
  - url: http://localhost:8003
    description: Node 3
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LockAcquireRequest:
      type: object
      required: [resource, owner]
      properties:
        resource:
          type: string
          example: "database-shard-1"
          description: Identifier resource yang ingin di-lock
        lock_type:
          type: string
          enum: [shared, exclusive]
          default: exclusive
          description: "shared = multiple readers; exclusive = single writer"
        owner:
          type: string
          example: "process-A"
          description: Identifier pemilik lock
        timeout:
          type: number
          default: 30.0
          description: Auto-release timeout dalam detik
        wait:
          type: boolean
          default: true
          description: Apakah menunggu jika lock tidak tersedia
    LockResponse:
      type: object
      properties:
        status:
          type: string
          enum: [granted, waiting, denied, deadlock]
        lock_id:
          type: string
        resource:
          type: string
        lock_type:
          type: string
    ProduceRequest:
      type: object
      required: [queue_name, payload]
      properties:
        queue_name:
          type: string
          example: "orders"
        payload:
          type: object
        producer_id:
          type: string
          default: "default"
    CacheWriteRequest:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          example: "user:123"
        value:
          description: "Any JSON value"
...
paths:
...
  /lock/acquire:
    post:
      summary: Acquire Distributed Lock
      tags: [Lock Manager]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LockAcquireRequest"
            examples:
              exclusive:
                value:
                  resource: "order-service-db"
                  lock_type: "exclusive"
                  owner: "worker-1"
                  timeout: 30
                  wait: false
              shared:
                value:
                  resource: "config-file"
                  lock_type: "shared"
                  owner: "reader-A"
      responses:
        "200":
          description: Lock acquisition result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockResponse"
        "403":
          description: Permission denied
  /lock/release:
    post:
      summary: Release Distributed Lock
      tags: [Lock Manager]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lock_id, owner]
              properties:
                lock_id:
                  type: string
                owner:
                  type: string
      responses:
        "200":
          description: Lock released
...
  /queue/produce:
    post:
      summary: Produce Message
      tags: [Queue]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProduceRequest"
            example:
              queue_name: "orders"
              payload:
                order_id: "ORD-001"
                amount: 150.00
              producer_id: "shop-service"
      responses:
        "200":
          description: Message produced
          content:
            application/json:
              example:
                message_id: "550e8400-e29b-41d4-a716-446655440000"
                queue_name: "orders"
  /queue/consume:
    get:
      summary: Consume Message (blocking)
      tags: [Queue]
      security:
        - BearerAuth: []
      parameters:
        - name: queue_name
          in: query
          required: true
          schema:
            type: string
        - name: consumer_id
          in: query
          schema:
            type: string
            default: consumer1
        - name: timeout
          in: query
          schema:
            type: number
            default: 5.0
      responses:
        "200":
          description: Message (null if timeout)
...
  /queue/stats:
    get:
      summary: Queue Statistics
      tags: [Queue]
      parameters:
        - name: queue_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Queue stats including pending count and responsible node
...
  /cache/write:
    post:
      summary: Write to MESI Cache
      tags: [Cache (MESI)]
      description: |
        Writes a value and broadcasts INVALIDATE to all peer caches.
        The writing node's state becomes Modified (M).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CacheWriteRequest"
      responses:
        "200":
          description: Write successful
          content:
            application/json:
              example:
                key: "user:123"
                success: true
                state: M
...